requirements.txt
# Core
python-dotenv>=1.0.0

# Voice Processing
SpeechRecognition>=3.8.1
pyttsx3>=2.90

# Web and API
requests>=2.28.1
beautifulsoup4>=4.11.1

# GPU and Deep Learning
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
--extra-index-url https://download.pytorch.org/whl/cu118

gpustat>=1.0.0
numpy>=1.21.0

# Utilities
python-dateutil>=2.8.2
pytz>=2022.1
